feat: introduce maintained single-symbol pipeline facade
